Create Automate-365-Licensing
This script automates the application of licensing using group membership. Whilst licenses can be applied to groups, and this is inherited to child groups, there are scenarios where
this doesn't work - in this particular script Dynamics licensing couldn't be applied simply at group level, as this would lead to users having licenses that duplicated function and therefore wasted money.
Some licenses are mutually exclusive; but not all of them. The logic in this script is specific to my organisation but can be adapted to suit.